2nd Commit after Udaciti'es Review
The parameters are set to reasonable numbers.
Nice work, but I recommend using an RNN size of 256 or 512 to let your model generalize better.
The main issue here is your embedding size. You need to increase it to give your room more room to train, otherwise, you're trying to compress information in too small of a dimensional space. I recommend 128.